avoid overwriting newly synced header info
In commit "a31e632 save move header to internal storage" there is a bug
where onStart() we regardlessly overwrite the AmbitRecord from storage,
which is the last saved headers. (saved in onStop())

In this use case:
1. user opens the app, app onStart() reads previously saved headers and
displays on screen
2. user puts the app in background, onStop() saves the same header to
storage.
3. now user attaches Ambit device, the app starts syncing headers
4. now user switches the app to foreground, onStart() executes and
overwrites the newly sync'ed headers by the old ones.

So this commit assumes that onStart() should only read saved headers
when there is no headers (log count == 0) in memory.