make a list of LogEntry locally for listview
Make a copy list of LogEntry from AmbitRecord every time we change it:
1. when syncing headers
2. when syncing a move
3. when onStart()

This copy list is for listview adapter so later we can call
notifyDataChanged() to update UI.

Because in LogAsyncTask and LogHeaderAsyncTask we modify the list of
LogEntry in the background thread. If we set this list to the list
adapter, AndroidRuntime will sometimes complain and the app crashes.