make GPX output "with barometer"
Append "with barometer" in the creator field of the GPX file so that
Strava will use the altitude field from the GPX instead of using
elevation map. This only makes sense if the Ambit device comes with
barometer. But if it does not, will it output elevation(altitude)
samples in the move?