store strava token in app pref and handle errors
We now store the strava token in app's shared preference so we can
re-use it until it expires. When it expires, user should re-login Strava
to get the new token.

Upload is working and some error handling is added. However, I still
need to work out how to implement a better UI in addition to that orange
strava button.